+++ CHANGELOG.md
@@ -14,6 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   outgoing trust sharepoint link field.
 - Fix statistics page to show accurate data for Transfer's `by region` table.
 
+### Added
+
+- Add the ability to download Transfer projects for the Grant management and
+  finance team
+
+### Changed
+
+- Changed the existing Grant management and finance team Conversions download
+  page to include an information page about the download, and a tabbed view on
+  the index page to switch between Transfers and Conversions
+
 ## [Release-49][release-49]
 
 ### Changed